feat: Add input form in picture page and some png files in public/images directory
