+++ frontend/src/components/pages/Picture.tsx
@@ -1,14 +1,32 @@
 import React from "react";
-import { SubmitButton, PictureButton } from "../index";
-import Grid from "@material-ui/core/Grid";
+import { PictureButton, InputField } from "../index";
+import { Button, makeStyles } from "@material-ui/core";
+
+const useStyles = makeStyles({
+    root: {
+        width: "80%",
+        fontSize: "1.5rem",
+        fontWeight: "bold",
+        color: "#ffffff",
+        background: "linear-gradient(45deg, #00d6af 30%, #00c3cf 90%)"
+    }
+});
 
 const Picture = () => {
+    const classes = useStyles();
     return (
-        <Grid item spacing={1} direction="column" alignContent="center" alignItems="center">
+        <>
             <div>여기에 바코드 등록</div>
             <PictureButton />
-            <SubmitButton />
-        </Grid>
+            <div>
+                <InputField text="Name" hint="Food Ingredient" />
+                <InputField text="Expiration Date" type="date" hint="Expiration Date" />
+                <InputField text="Cateogry" hint="Ingredient Category" />
+            </div>
+            <Button variant="contained" color="primary" className={classes.root}>
+                Save
+            </Button>
+        </>
     );
 };
